refactored Github port into just use cases
