+++ Web Ui/src/modules/Auth/application/signOutWithGithub.ts
@@ -1,0 +1,11 @@
+import { getAuth, signOut } from "firebase/auth";
+import firebase from "../../../firebaseConfig";
+
+export async function handleGithubSignOut() {
+  const auth = getAuth(firebase);
+  try {
+    await signOut(auth);
+  } catch (error) {
+    console.error("Error al cerrar sesión", error);
+  }
+}
