+++ Web Ui/src/sections/Invitation/InvitationPage.tsx
@@ -4,7 +4,6 @@ import firebase from "../../firebaseConfig";
 import SuccessfulEnrollmentPopUp from "./components/SuccessfulEnrollmentPopUp";
 import Button from "@mui/material/Button";
 import GitHubIcon from "@mui/icons-material/GitHub";
-import { GithubAuthPort } from "../../modules/Auth/application/GithubAuthPort";
 import { RegisterPort } from "../../modules/Auth/application/RegisterPort";
 import UserOnDb from "../../modules/Auth/domain/userOnDb.interface";
 import Card from "@mui/material/Card";
@@ -12,10 +11,11 @@ import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
 import Typography from "@mui/material/Typography";
 import { Grid } from "@mui/material";
+import { handleSignInWithGitHub } from "../../modules/Auth/application/signInWithGithub";
+import { handleGithubSignOut } from "../../modules/Auth/application/signOutWithGithub";
 
 function InvitationPage() {
   const [user, setUser] = useState<User | null>(null);
-  const githubAuthPort = new GithubAuthPort();
   const dbAuthPort = new RegisterPort();
   useEffect(() => {
     const auth = getAuth(firebase);
@@ -30,7 +30,7 @@ function InvitationPage() {
   const [showPopUp, setShowPopUp] = useState(false);
 
   const handleSignUp = async () => {
-    let userData = await githubAuthPort.handleSignInWithGitHub();
+    let userData = await handleSignInWithGitHub();
     if (userData) {
       setUser(userData);
     }
@@ -108,7 +108,7 @@ function InvitationPage() {
                       </Grid>
                       <Grid item sx={{ marginTop: "auto" }}>
                         <Button
-                          onClick={githubAuthPort.handleSignOut}
+                          onClick={handleGithubSignOut}
                           variant="contained"
                           color="primary"
                         >
